+++ app/helpers/alchemy/url_helper.rb
@@ -43,7 +43,7 @@ module Alchemy
       url_params = {
         id: picture.id,
         name: picture.urlname,
-        format: configuration(:image_output_format),
+        format: picture.default_render_format,
         sh: picture.security_token(optional_params)
       }
       url_params.update(optional_params.update(crop: optional_params[:crop] ? 'crop' : nil))