+++ config/initializers/mime_types.rb
@@ -1,0 +1,1 @@
+Mime::Type.register "image/svg+xml", :svg